+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Make renamed copes of individual `ai_config.json`
 - Import personalities as server scene mod
+- Install task to overwrite `commander_manager`
 - Update commander names with AI name
 - Use default commanders
 - Typo correction in Queller